parse xml devices
There are lots of devices can be configured for guest, here only list import ones.

<devices>
  <hostdev mode="subsystem" type="pci" managed="yes">
    <source>
      <address type="pci" domain="0x0000" bus="0x04" slot="0x00" function="0x0"/>
    </source>
  </hostdev>
</devices>

[hostdev detail](https://libvirt.org/docs/libvirt-appdev-guide-python/en-US/html/libvirt_application_development_guide_using_python-Guest_Domains-Device_Config-PCI_Pass.html)
[hostdev parameter](https://www.ibm.com/docs/en/linux-on-systems?topic=xml-hostdev)

hostdev(node)[pci passthrough] is pci devices from host, it's physical device, it's also bound to driver(usually) and insert to host system, hence you can see it
from host, but in order to let guest use this, you have to detach the device from host, then assign it to guest, so that guest starts and detects the devices, bind it with virtio-pci, like above managed=yes,
libvirtd will detach it and reattach it from host automatically depends on guest starts or shutdown, if no, you have to do this by yourself.

NOTE: above is for standard PCI device, for devices which support SR-IOV(VFS) most for network and block devices, use "intelligent passthrough" of network devices like below, only valid for SR-IOV

<devices>
  <interface type='hostdev' managed='yes'>
    <driver name='vfio'/>
    <source>
      <address type='pci' domain='0x0000' bus='0x00' slot='0x07' function='0x0'/>
    </source>
    <mac address='52:54:00:6d:90:02'/>
    <virtualport type='802.1Qbh'>
      <parameters profileid='finance'/>
    </virtualport>
  </interface>
</devices>

it is very similar to the functionality of a standard <hostdev> device, the difference being that this method **allows specifying a MAC address and <virtualport> for the passed-through device**. If these capabilities are not required, if you have a standard single-port PCI, PCIe, or USB network card that doesn't support SR-IOV (and hence would anyway lose the configured MAC address during reset after being assigned to the guest domain), you should use standard <hostdev> to assign the device to the guest instead of <interface type='hostdev'/>